Revert "add trace API"
This reverts commit 3e66dca89b4c0cf8d74e5a66ae2095fda94251a0.
