--- pkg/util/filesystem/go.mod
@@ -1,25 +1,0 @@
-module github.com/DataDog/datadog-agent/pkg/util/filesystem
-
-go 1.19
-
-require (
-	github.com/DataDog/datadog-agent/pkg/util/log v0.43.1
-	github.com/hectane/go-acl v0.0.0-20230122075934-ca0b05cb1adb
-	github.com/shirou/gopsutil/v3 v3.23.3
-	github.com/stretchr/testify v1.8.2
-	golang.org/x/sys v0.7.0
-	gotest.tools v2.2.0+incompatible
-)
-
-require (
-	github.com/DataDog/datadog-agent/pkg/util/scrubber v0.43.1 // indirect
-	github.com/cihub/seelog v0.0.0-20170130134532-f561c5e57575 // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/go-ole/go-ole v1.2.6 // indirect
-	github.com/google/go-cmp v0.5.9 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/power-devops/perfstat v0.0.0-20210106213030-5aafc221ea8c // indirect
-	github.com/yusufpapurcu/wmi v1.2.2 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
