--- pkg/util/net/go.mod
@@ -1,3 +1,0 @@
-module github.com/DataDog/datadog-agent/pkg/util/net
-
-go 1.19
